Define a way to dynamically provide vlaidationRules
